optimize: move GoogleService-Info.plist with switch-env.sh

+++ switch-env.sh
@@ -12,7 +12,7 @@ if [ "$ENV" != "dev" ] && [ "$ENV" != "sit" ] && [ "$ENV" != "prod" ]; then
     exit 1
 fi
 
-# Check if files exist
+# Check if the Android files for the specified environment exist
 if [ ! -f env/$ENV/android/strings.xml ]; then
     echo "Error: File env/$ENV/android/strings.xml does not exist!"
     exit 1
@@ -29,11 +29,20 @@ if [ ! -f env/$ENV/ios/Info.plist ]; then
     exit 1
 fi
 
+# Check if the GoogleService-Info.plist file for the specified environment exists
+if [ ! -f env/$ENV/ios/GoogleService-Info.plist ]; then
+    echo "Error: File env/$ENV/ios/GoogleService-Info.plist does not exist!"
+    exit 1
+fi
+
 echo "Switching to environment '$ENV'..."
 
-# copy the files
+# Copy the Android files
 cp env/$ENV/android/strings.xml android/app/src/main/res/values/strings.xml
 cp env/$ENV/android/google-services.json android/app/google-services.json
+
+# Copy the iOS files
 cp env/$ENV/ios/Info.plist ios/App/App/Info.plist
+cp env/$ENV/ios/GoogleService-Info.plist ios/App/App/GoogleService-Info.plist
 
 echo "Environment '$ENV' set successfully."
